Update top page and css
Need some custom css for eSolia logo, and in a couple places.

The materialize.css base is good, but, it does not work well for certain things, so you need to adjust a little. For instance making a skinny horizontal bar which is colored all the way across (hint: needs to be put in another div to color across).