Add valign to top nav and set media queries
Removed padding from logo class, and added valign-wrapper and valign to
top nav, so that the logo will naturally be vertically aligned.

At the same time, something is setting the height of the nav area to
64px when the screen is <600px, so used media query in style.css to
check for that size, and make the logo a little smaller.

Tested, and logo resizes as expected.