Add generic opengraph code for node type pages
Node type pages, like the home page, are going to need a different sort
of og code, compared to a content page. Started work on this, adding
a data file to hold URLs of generic site images, and referenced these for
a node, where there would not necessarily be any images available to
pick from.