Add "why esolia" and "features" sections to home
Used the grid to do a kind of bullet list which will be responsive, for
the why esolia piece. First tried bright bullets, but it's distracting,
so went with more subdued indigo and brown.

Learn more links, as well as many other links have to be set before go-
 live.

Using "card" feature from materialcss to make the features, and will add
RSS feed from a taxo term, for the Latest list.