Add second rss feed pull for a taxonomy
Hugo makes an rss feed for all taxonomies, so, we can use this to our
advantage. We can pull the rss for any taxo into the home page, via
a javascript, and have a separate list of posts, for instance, a list of
featured products of services.

Also, we'll use the static cards for special "pinned" features, and do
those manually for now. It is probably possible to pull them dynamically
via hugo, though.

Fixes #3.