Adjust header balance for Noto jp font
As @eishikawa pointed out, the headers were too large. With the
original Roboto font (used in Android and some Google web properties)
they look ok, but, with Noto Japanese, which has a high X height, they
look fat. So adjusted the balance of the headers, so the h1 is 3.5rem
and the h6 is 1.2rem.

Fixes #13