Pull post-checkout into root of repo
If the git hook scripts are in the root of the repo, then you can make a
 symlink in .git/hooks _from_ the script in the root. When you do this,
the script is tracked, and, it will run for everyone who clones the repo.

That said, this particular post-checkout script is kind of faulty, since
it runs for every `git checkout` and not selectively.