Add fixed part of side menu to config_xx.toml
If you're iterating over a menu in hugo templates, you can also pull in
menus from the config.toml file. Set this up so zerobin is a child of a
"utilities" parent, in both configs, and tested. Works fine.

Sidebar.html had a bug in that it was not referring to side but rather
to main menu. Also, I used {{ .Post }} to specify target=_blank for the
off site links.