Add new titlewbr field
Using the <wbr> and <nobr> tags is all well and good, but
if you use them in a title, since only one title field
was being used for everything, the "raw" tags were appearing
in the browser tag, or in social posts.

To fix, made another param called titlewbr, and _if_ these
are needed, make a copy of the title tag, change the key to
titlewbr, and add the tags there.

The detailpage-header partial will check for the existence
of this tag, and use it for _display_ on the page, but,
the title will be as-is.

Fixes #154